Rename ASTParser#parse_expression to #parse_instructions
This is a private method, so it doesn’t really matter what we call it,
but by _not_ calling it #parse_expression we free up that name for reuse
by the public method currently called #parse. The point of this is to
make space for a second public method called #parse_script.