Move Interpreter#define_function to ASTParser#parse_function
This was already basically a parser method, so it just needed moving
over to the right place.

It’s written in the wrong style — instead of repeatedly destructuring an
array into its head and tail, it should be storing the S-expression in
the parser’s state and using the helpers to destructively consume it.
I’m going to wait until all the parsing methods are extracted from the
interpreter before I convert all of them to this style.