Extract ASTParser#parse_memory from the interpreter
The extracted code in #parse_memory was the only client of #parse_string
and #interpret_integer, so they can be removed from the interpreter.
I’ve adapted #parse_string to look for its string in the parser state,
even though that requires some temporary fixing up in #parse_memory
because the latter doesn’t use the parser state yet.