Stop returning enumerator from SExpressionParser#parse
We can’t make use of this at the moment, because the AST parser expects
to get an array. It’s a shame to lose the incremental parsing, but
there’s really no point in getting an enumerator from this method
unless/until we can make actual use of it rather than just immediately
flattening it out into an array.

I may well want this behaviour back at some point, but if so, I can just
revert this commit. Until then I won’t need to think about whether I’m
dealing with an array or an enumerator.

This reverts 2f4d7d3fcb6546d02bb94308a671658743dc0ef4 and
1721608e76ab3e81599bb3a69c5356f22d8da591.