Move `terminated_by:` outside of #parse_instructions
This means that the caller needs to compose #with_input and #read_until
if they want to control how much of the input can be read. But that’s
good, because it means we can use the same compositional strategy
everywhere, including for different parser methods, rather than having
the #parse_instructions method be somehow special. (It isn’t.)