During function invocation, dereference type index if present
While we’re still handling names in the interpreter (i.e. until I
finally implement the identifier context in the AST parser), we need to
be careful that the names from the type definition don’t leak through
into the local variable names, because the spec says they’re purely for
documentation purposes. [0]

[0] https://webassembly.github.io/spec/core/text/types.html#function-types