Add block comment support to S-expression parser
This regular expression is pretty nasty, but it corresponds directly to
the grammar in the spec [0], and I’m having difficulty thinking of a
nicer way of doing it off the top of my head. I’m sure there is one, but
when I think of it I’ll just come back and change it.

Alternatively, I don’t actually have to do all this with a single
regular expression — this is a recursive descent parser, after all, so I
could just implement #parse_nested_comment which can call itself
recursively rather than rely upon a subexpression call in the regexp.
I’m going to stick with the regexp for now because it’s working fine and
I don’t really care about the S-expression parser any more.

[0] https://webassembly.github.io/spec/core/text/lexical.html#comments