Rename ASTParser#with_input to #read_list
With a bit of distance from when I first named this method, it does
appear to be used only when we want to read a complete list from the
input.

For my own safety I should probably be asserting that a) the thing I’m
trying to read is actually a list (vs an atom), and b) the list I’ve
tried to read has been read to completion. I’ll do that next.