Remove #read_list argument when it’s just `read`
There are a few cases where either we’re providing something that’s not
the same as `read` (e.g. in #parse_expression where we’ve received a new
S-expression as an argument), or we’re providing `read` but we need to
keep it in a local variable for some weird reason (e.g. `expression` in
the #parse_function method). It might be possible to refactor around the
latter, but I’ve taken care of all the obvious cases for now.