Get rid of misleading `begin`/`ensure`/`end` block in #read_list
This makes it look like there’s something exotic happening to do with
cleaning up after exceptions, but it’s not that at all — I was just
using `ensure` as a lazy way to avoid having to store the result of
`yield` in a local variable (and I didn’t want to use #tap for some
reason). Given that I’m making an effort to be clear rather than clever,
I think it’s preferable to use the dead-simple approach of just
assigning to a local variable.