Get rid of Script AST node
At the moment an expression is represented everywhere as just an array
of instructions, so for now let’s allow a script to be represented as
just an array of commands. If this turns out to be a mistake for any
regularity-related reason, we can easily change it back the other way
and make both things AST nodes. But at least this way it’s both simple
and consistent.