Extract ASTParser#repeatedly helper
It turns out we’re repeating this pattern all over the place, so it’s
helpful to wrap it up and give it a name.

In some cases we don’t care about the result of the block and we just
throw the accumulated array away; I’m comfortable with the performance
cost of doing that, since it makes the code clearer than having two
separate helpers, one which collects the values and one which doesn’t.