Only read lists from function body when we know they’re part of its signature
This is fairly easy now that we have #can_read_list?(starting_with:).

I find it incredibly satisfying that we can now say `body =
parse_instructions`. This is possible because we only need to repeatedly
read the lists in the signature, and as soon as we can tell we’ve
finished doing that, the entire rest of the current S-expression is just
the body. This is so much nicer than what we were doing before to piece
together the parts of the body before parsing it. I’m extremely happy
that it works.