Use #repeatedly to read minimum and maximum memory size
This makes it clearer that `memory` is followed either by `(data …)` or
many integers — actually only one or two. By trying to parse everything
after `memory` as an integer in this case, we’ll fail faster if there
are more than two values, or (in principle if not in practice given how
tolerant #to_i is) if any of them is not an integer.