Extract #parse_memory_data helper
The convention is for #parse_* methods to begin by reading their own
opcode. In order to do this here, we need to arrange for the called to
use #read_list without a `starting_with:` argument, so that the opcode
is left alone for #parse_memory_data to consume.

I’m not necessarily confident that this is the right design for #parse_*
methods, but it’s how they all work at the moment, so I’m trying to be
consistent if nothing else.