Read an ID, not an index, in #parse_table
This was just a mistake: `table` can only be followed by an ID. This
incorrect regular expression was hiding the fact that we’d overlooked
the possibility of `table` being followed by numeric limits. Now that
we’re explicitly parsing (and discarding) these, we can be more strict
about the syntax of the table name.