Extract #parse_typeuse helper
The `typeuse` [0] includes the type index, the `param`s and the `result`s,
so we can parse them all at once.

Ultimately a function should be represented in the abstract syntax by
only a type index, a vector of locals and a body [1], e.g. by generating
a new typedef for the function’s type if a suitable one doesn’t already
exist, but we won’t be ready to tackle that until we have an identifier
context.

[0] https://webassembly.github.io/spec/core/text/modules.html#text-typeuse
[1] https://webassembly.github.io/spec/core/syntax/modules.html#functions