Thread a dummy identifier context through to #parse_normal_instruction
We want to use the identifier context to look up the index of a global,
so we need the context to reach all the way from #parse_text_fields
(where we’re expecting to compute the initial context [0]) to
the #parse_normal_instruction method, where we parse the `global.get`
and `global.set` instructions.

Passing the identifier context around is already inconvenient enough
that we’ll probably want to store it as state on the parser object
instead (as we do with the current S-expression), but I’d like to get
something simple working first.

In #parse_invoke and #parse_expecteds I’ve chosen to provide an empty
context because the relevant commands only appear outside of modules,
and hence shouldn’t have any identifiers in scope.

[0] https://webassembly.github.io/spec/core/text/modules.html#text-module