Build initial identifier context from a copy of the module fields
Building the whole context up front makes it possible to deal with
forward references, e.g. calling a function which is defined later in
the module.

It’s pretty gross to use Marshal to take a deep copy of the
S-expression, but it’s a quick and dirty way to get something working
without destroying the original S-expression.

Once this is all working, I’d like to switch back to a single pass over
the S-expression rather than processing it twice and taking a deep copy.
I’d do this by making the various #parse_* methods return an updated
identifier context (if they parsed something that affects the context)
and a proc which can be called with an identifier context in order to
produce an AST node. This will allow us to collect up the final
identifier context and then pass it into all those procs to get the
final abstract syntax tree. But that’s a bit complicated, so I’d rather
build up my confidence first by getting everything working and save that
approach for a final refactor.