Automatically insert a type index referring to the first matching type definition
This can still be `nil` at the moment if there is no such matching
definition, but that doesn’t cause any problems yet, because we’re not
trying to use that type index to look up the appropriate type definition
in the identifier context. Before we can start doing that, we need to
automatically insert missing type definitions where necessary, and I’m
going to implement that next.