Replace #with_branch_handler `jump:` argument with optional `redo_on_branch:`
@zetter used an optional boolean argument in #2 and I really like that
idea, not least because it means we only need to provide the extra
argument in the one case where we want `redo` to happen. He called the
argument `branch_to_sequence_start:` in his PR, but I’ve gone with
`redo_on_branch:` because it’s the most simple-minded name I could
possibly think of, which means it’s more likely I’ll remember what it
does in future.