Parse typeuse lookup into #parse_typeuse
We’re currently calling #parse_typeuse in #parse_function, but it turns
out we need to do the same thing when parsing structured instructions,
because those can also contain a typeuse. So rather than duplicate all
of the code that’s responsible for looking up the type index when it’s
not provided (including potentially generating a new type definition for
the index to point to), it’s better to just move that code inside
the #parse_typeuse method so that any other caller can also get the same
functionality.

I’m currently very apprehensive about the specification of typeuses [0],
which suggests that the resulting identifier context is empty except for
the (possibly blank) names of the locals generated by the typeuse, and I
can’t see where the original identifier context actually gets reused.
For now I’m just rolling with my own interpretation of the prose, i.e.
the resulting identifier context is an _updated_ version of the original
one, not a new blank one. I don’t know whether that’s right but it seems
necessary to keep things working correctly for now.

This method, and its return values, feel very messy. I’m hoping that
some of this mess goes away when we’re no longer doing a two-pass
operation to produce an AST, but I don’t really have anything to base
that on.

[0] https://webassembly.github.io/spec/core/text/modules.html#type-uses