Use type definitions from updated context instead of reconstructing them
Every `type` in the module has already been parsed during
the #build_initial_context method, which is how we’re able to correctly
compute the indexes of generated types inside #parse_typeuse in the
first place, so it’s unnecessary work to recompute this list of types by
re-parsing the `type` fields during the second pass and accumulating any
generated types from the function definitions. We can use the typedefs
from the updated identifier context directly instead.

This commit will allow us to remove a bunch of other code (e.g. stop
returning and accumulating the generated types), but this is the
smallest meaningful change.