Add `context` as parser state, and #with_context to manage it
To correctly support block types (which may generate new type
definitions) we need to be able to update the current identifier context
in many more places than we currently do (i.e. function definitions),
which is going to make it extremely hard to correctly thread the context
through every operation.

So, instead, let’s store the current context as state on the parser, and
have a #with_context method which allows us to temporarily reassign it.
This’ll make it much easier for arbitrary methods to update the current
context when they need to, without forcing their caller to deal with a
new context as a return value.

I don’t feel great about doing this, but I think the alternative (i.e.
tons of boilerplate code to thread the current context around
everywhere) is even worse, so this approach is almost certainly the
lesser evil.