Replace `results` with `type` on structured instructions
In general, blocks can have parameters as well as results, so we need to
store a function type for the block, not just a result type. At the
moment we’re only populating that function type with the result type,
but we’re trying to get to the point where we can support block
parameters, so we’ll need the rest of the function type soon.