Parse block type with #parse_typeuse instead of #parse_results
This gives us the ability to parse blocks which have parameters as well
as results.

The spec says that block types should behave differently in “the special
case of a type use that is syntactically empty or consists of only a
single result” [0]. We haven’t addressed that special case yet.

We also haven’t addressed the fact that we’re meant to store a type
index (or, in the aforementioned special case, an optional value type),
not the type itself, as the block type; for this commit I’m trying to
change the parser without breaking compatibility with the interpreter.

This does mean we’re now potentially modifying the identifier context
when we parse a function body, so I’ve added a TODO comment to note that
we shouldn’t just throw it away when we’ve finished parsing the function
definition. At the moment this doesn’t appear to cause any test
failures, so I’m not going to try to fix it yet.

[0] https://webassembly.github.io/spec/core/text/instructions.html#text-blocktype