Don’t discard new typedefs generated by parsing function body
This was causing type indexes to (potentially, when we start doing it)
be computed incorrectly in float_exprs.wast, because the block types in
it were being discarded at the end of the function body, and that left
us with type indexes that were pointing outside of the final list of
typedefs.

I don’t feel good about the way I’ve solved this, because we probably
shouldn’t be hacking about with the identifier context like this, but it
gets the job done for now and allows us to make progress. I’d like to
come back later and think harder about how to better organise the
information which we accumulate during parsing so that we don’t have to
weld together two different identifier contexts like this.