Support the blocktype special case of an empty or single result type
The spec says [0] that a block which has syntactically zero or one
results, and no parameters, should be treated specially: we provide that
result type inline in the AST, and never add a new type definition. To
support this I’ve duplicated the hack from #parse_function and used the
old & new typedef lists to decide whether a typedef has been generated
of this specific shape, and return an inline result type (discarding the
updated identifier context) if so.

Again I’m not very pleased with this code but it does appear to work.

[0] https://webassembly.github.io/spec/core/text/instructions.html#text-blocktype