Replace `type_index` with `type` again in Block, Loop and If
This essentially reverts cd48ab90e3f10803ca15fe7dfd54b38337b7fbed, which
didn’t account for the block type special case. Now we do account for
it, and `type_index` is a misleading name because the value stored in
the AST is no longer necessarily a type index: it could also be an
inline result type.