Correctly account for block parameter arity in branch handler
When restoring the stack after jumping out of a block, we need to
account for the values that were consumed by the block. At the moment
it’s only the (unimplemented) validation phase which checks that the
block really does consumes them, but on the assumption that the program
is valid, we know that the stack will have fewer items on it after the
block completes if the block has a non-zero number of parameters.