Rename `saved_operands` local variable to `result_values`
This is actually what they are, which will become clearer when we only
push them onto the stack in the case of a “forward jump” [0], i.e. a
branch out of a `block` or `if`, and not in the case of a “backward
jump”, i.e. a branch out of a `loop`.

[0] https://webassembly.github.io/spec/core/syntax/instructions.html#control-instructions