Restore the correct number of values when unwinding the stack after a backwards jump
The specification says [0] that backward jumps (i.e. branches inside a
`loop`) should pop off the number of values corresponding to the input
type of the loop body, and restore them after unwinding the stack.
Previously we were totally ignoring this and always using the number of
values corresponding to the block’s output type, which is why the
`loop.wast` tests were failing. They’re all passing now!

[0] https://webassembly.github.io/spec/core/syntax/instructions.html#control-instructions