Use Data#with in #with_sign implementations on floats
We still need #with_sign so that it can be used polymorphically (i.e.
Wasminna::Float::Finite doesn’t have a `sign` member), but the
implementation is now much nicer because we can use Data#with to copy
all of the other members instead of having to call #to_h.