Exclude label index space from identifier context well-formedness check
I’m not happy about this, because the spec [0] strongly implies that
we’re not allowed to put duplicate labels into the identifier context,
but some of the test cases (e.g. the `redefinition` function in
labels.wast) definitely have nested blocks which shadow the labels of
their enclosing blocks, so it looks like we have to support it.

I expect I’ve misunderstood something about the spec, but despite a few
minutes of exploring, I can’t work out what that is. It’s not impossible
that the spec is wrong about this; it’s just much more likely that I’ve
misunderstood it.

[0] https://webassembly.github.io/spec/core/text/instructions.html#labels