Add new labels to the identifier context as we discover them
Per the specification [0], it’s important that the new labels are
prepended to the existing labels in the identifier context, so that the
indexes of all existing labels are incremented. It’s also important that
we do this even when the label is not present (i.e. it’s `nil`) so that
those indexes are still incremented.

I haven’t done anything about the side condition stating that the label
can’t already appear in the identifier context, because that doesn’t
seem to true in practice per the previous commit. Again, it’s likely
I’ve just missed something here, but it definitely doesn’t look like it
should be a syntax error, because it’s something the official tests do.
Shrug.

[0] https://webassembly.github.io/spec/core/text/instructions.html#labels