Look up labels in identifier context during parsing
This means that we’re no longer storing strings in the AST nodes for the
`br`, `br_if` and `br_table` instructions; the target of a branch is now
always represented in the AST with a number, never a symbolic identifier.

In the next commit we’ll be able to remove the interpreter code which
handles symbolic identifiers as branch targets.