Extract storage size and sign extension mode from load operation name
Load instructions (e.g. `i32.load`, `i32.load8_s`, `i64.load32_u`) can
contain a storage size and a sign extension mode [0] for use in cases
where the value to be loaded is less wide than the result type. At the
moment we’re completely ignoring this, but it obviously affects the
behaviour of the instruction, so we need to parse it out of the
operation name and store it in the AST; we’ll do the latter part in the
next commit.

I considered using a boolean (e.g. `signed`) instead of a symbolic sign
extension mode, but ultimately it didn’t feel as though I had a good
reason (beyond “there are two options”) to use a boolean for this.
Neither mode is more important than the other; they’re just two
different modes of operation for the instruction. Ultimately it felt
clearer and more obvious to have an explicit `sign_extension_mode`
member for the AST node.

[0] https://webassembly.github.io/spec/core/syntax/instructions.html#memory-instructions