Move stack height and branch arity calculations out of block in #with_branch_handler
I never intended for these lines of code to be inside the block (which
gets redone if it’s a loop body), and it’s fortunate that the code
worked even with that unintentional design. It makes more sense for the
stack height and branch arity to be computed once at the start and not
repeated for each potential loop iteration, so let’s do that instead.