Use a separate boolean to detect whether a branch has happened
Previously we were using in-band information from `result` to decide
whether the `catch` block had completed successfully or not, which
worked fine (for now) but was a bit messy and hard to understand.
Instead, we’re now assigning `false` to an initially-`true` local
variable if the block completes successfully, which I think is easier to
understand, and also happens to be robust against pathological cases
like `throw(:branch, :did_not_throw)`.