Extract #evaluate_block helper
This wraps up the pattern of expanding a blocktype and then evaluating
the block’s body with a branch handler. I’m keeping this separate for
now rather than integrating it directly into #as_block, but if it turns
out to be the only caller of #as_block it might make sense to combine
them; likewise for #expand_blocktype. (On the other hand, there’s
something to be said for keeping methods short and focused, even if that
means there’s a lot more of them.)