Remove `redo_on_branch` default from #as_block
I chose to give this a default in #evaluate_block, which is just
repeating the default from #as_block, and I don’t want it in both
places, so I’m removing it from here.

Overall I’m not happy about the name (or even the existence) of this
argument, but right now I don’t have any brilliant ideas about what to
do instead. If any brilliant ideas occur to me in future, I’ll come back
to it.