Move `return` handler into #invoke_function
This is a more natural place to unwind the stack if necessary, because
it mirrors the way we handle stack unwinding for branches. As a bonus, I
think we no longer need #with_current_function because this code now
lives in the same lexical scope as the information it needs (i.e. what
function we’re returning from), which is a pretty strong signal that
this is a better place for it to live.