Remove unnecessary use of #with_current_function
The previous commit moved the `return` handler to a place where there
was already a `function` local variable in scope, so the previous call
to the #function getter was no longer being used. The only purpose of
the #with_current_function method was to update the value of that
attribute, so we no longer need to call it here.