Use #as_block to unwind the stack upon `return`
The #invoke_function method was repeating the implementation of stack
unwinding: it was remembering the original height of the stack (minus
the function arguments) and then restoring that height if a `return`
instruction executed. We already have the code to unwind the stack in
the #as_block method, and we can reuse it here by simulating a branch
whenever a `return` occurs.

In a way this is slightly hacky, but in another way it corresponds
closely to the specification [0], which says:

> The `return` instruction is a shortcut for an unconditional branch to
> the outermost block, which implicitly is the body of the current
> function.

By catching a `return` at the outermost block, and converting it into an
unconditional branch to the now-innermost (i.e. zeroth) block, we’re
providing exactly this shortcut. So that’s absolutely fine.

This new caller of the #as_block method is making me slightly regret the
introduction of #evaluate_block, because we now have a different way to
use #as_block, which undermines the earlier observation that it was
always being used in the same way. There’s also now a weird impedance
mismatch between type indexes and types — we look up the function type
twice to make it available in two different places. I can’t immediately
see a nice way to deal with these problems; maybe I’ll end up reverting
the creation of #evaluate_block altogether, but it’s still useful inside
the #invoke_function method so that’s not necessarily correct either.
Needs more thought.

[0] https://webassembly.github.io/spec/core/syntax/instructions.html#control-instructions