Consider labels index space when deciding identifier context well-formedness
The spec [0] doesn’t exempt the labels index space from its definition
of well-formedness, but I had exempted it because Context#+ needs to be
able to create identifier contexts containing duplicate labels due to
label shadowing. Now that we’re only checking well-formedness in a more
limited number of situations, it’s okay to consider the labels index
space again during that check.

[0] https://webassembly.github.io/spec/core/text/conventions.html#text-context-wf