Stop creating instances of Local AST node
These instances don’t contain any data and have no behaviour, so they’re
not doing anything useful except giving the `function.locals` array the
correct length, which is what the interpreter needs to know. We were
previously throwing away the type, but we might as well keep it and
store it in the array instead of a Local instance, since it does the
same job (of giving the locals array the correct length) and will
actually come in useful in future when we implement verification.

This has the nice bonus of making the #parse_local method almost
identical to #parse_parameter, so we could now combine them if we wanted
to.