Combine #parse_parameter and #parse_local
These methods are identical except for the starting atom they read, but
because it’s the caller’s responsibilty to make sure that these methods
are only called when the correct atom is next, we’re not losing anything
by combining them and being a little more permissive about the value of
that atom.

We can’t currently combine these with #parse_results because that has a
different return type and enforces the absence of a symbolic identifier.