Use generated tag objects for branching instead of `:branch` and a number
This allows the branch instructions to directly target one of the
enclosing blocks, by throwing the unique object corresponding to that
block, instead of throwing the generic `:branch` symbol and relying upon
an integer index payload to either handle or rethrow that symbol.

I think that this is a bit clearer and simpler than the previous setup,
although it does require us to maintain the stack of current tags so
that branch instructions can retrieve the appropriate one (based on the
index of the branch). The real benefit is going to be when we use these
tags to implement `return` — I’m expecting we’ll be able to reuse the
existing tags, which will unify the mechanisms of branches and returns —
but I haven’t done that yet.