Always execute function bodies with an empty tag stack
It should be impossible for instructions within a function body to ever
branch to blocks outside of that body, but right now a function call
will allow the called function to see tags from the caller. We can
prevent this by emptying the tag stack when we invoke a function, and
restore that stack once the the function invocation is over.

This change doesn’t make much sense on its own; it’s preparatory work
for changing the implementation of `return`.