Use `loop`/`break` instead of `tap`/`redo` in #as_block
Although this is to some extent undoing the good work of #2, I notice
that it puts the loop termination condition syntactically right at the
end of the loop, which should make it possible for us to move that
condition outside of the loop altogether, which is hopefully clearer
than the current `redo` situation.

I don’t feel particularly confident about this, and can only hope that
@zetter will forgive me if I turn out to be wrong. To be honest, my main
motivation is to just get rid of the unnecessary #tap, which upset me.