Use `begin`/`while` instead of `loop`/`break` in #as_block
We’re now able to move the termination condition outside of the loop
body entirely, which creates the tantalising possibility of being able
to separate these two concerns. I do miss the #redo but ultimately I’m
hopeful that this is going to be clearer in the end. We shall see!