Extract #with_branch_handler from #as_block
This is the part of the block-handling code which will be executed
either once (in the case of a normal block) or perhaps multiple times
(in the case of a loop) depending on what kind of block we’re executing.
By extracting this I’m hoping I can split the #as_block helper into two
separate ones dedicated to the two kinds of block, thereby eliminated
the cursed `redo_on_branch:` boolean parameter.