Move stack height measurement into #with_branch_handler
This happens to work fine, because either this is the first time the
block has executed, in which case it doesn’t matter in which method we
measure the initial stack depth, or it’s not the first time it’s
executed, in which case the branch handler from the previous iteration
(which is what must have caused it to execute again) has already
restored the stack to its original height.

This is slightly subtle, and potentially harder to understand than the
previous setup where we only ever measured the stack height once at the
beginning, but I think it’s worth doing it this way in order to minimise
the amount of setup code in #as_block, which we’d like to split out into
an accompanying #as_loop method. Overall I hope it’s a net win for
comprehensibility of this code, even if this specific step is a little
hard to justify in isolation.