Inline the now-unnecessary #evaluate_block helper
When I introduced this it was the only way to evaluate a block, but now
it’s just one of several ways (e.g. #as_loop, and the different usage of
the #as_block method inside #invoke_function), and in fact is only used
twice. So it’s no longer particularly special and therefore doesn’t
deserve to exist. Harsh but fair.

I was extremely tempted to push the call to #expand_blocktype inside
the #with_branch_handler method, since it looks like we’re always doing
it, but that’s not quite true: in #invoke_function we don’t want the
option of passing in an array of result types, only a type index. This
is quite pedantic but it doesn’t feel like a significant enough gain to
do the #expand_blocktype call somewhere else if that isn’t strictly
correct, so I’m going to chicken out of it for now.