Support parsing value type as global type in #parse_global
I’m not yet sure what this means semantically (e.g. does it imply that
the global is immutable?) but it’s legitimate to use a bare value type,
e.g. `i32`, instead of e.g. `(mut i32)` when declaring a global [0], so
we need to be able to parse it.

[0] https://webassembly.github.io/spec/core/text/types.html#global-types