Support reading multiple `(export …)` abbreviations in #parse_function
There can be arbitrarily many of these, so we need to be able to read
all of them, not just one. I’ve inlined #parse_export because it doesn’t
do the right thing any more, and it’s only called in this one place
anyway, so for now it’s clearer to have the logic inline.