Back out of returning an array of results from #parse_function
I’ve changed my mind about this, because it is (for now) going to be too
complicated to split out exports (and, in future, imports) into their
own independent AST nodes, due to the complexity of creating fresh
identifiers to match anonymous functions to their corresponding exports.

I can work around this by storing export (and, in future, import)
information on the function AST node itself, which is a little bit of a
hack, but is going to make my life easier in the short term. I might
revisit this decision once I’ve got single-pass parsing working, because
then it may become more feasible to create a fresh identifier in a
single place and reuse it between the export node and the function node.

This reverts commit 5b49b66d61d2f1a2afaed8137da8f7a78c4763cd.