Support multiple exported names for function definitions
As explained in the previous commit, we are (for now) going to overload
function definitions to also represent function exports. Because it’s
legitimate to export the same function multiple times within a single
definition, we need to be able to store all of these exported names as
part of the function’s (overloaded) AST node, not just the first or last
such name.