Use module exports as a fallback when looking up exported function names
It’s now the case that, somewhat unfortunately, exported function names
are available in two different places: inline exports [0] store names on
the function definition itself, whereas top-level exports [1] store
names and function indexes on the module. So we need to look in two
different places when trying to resolve an exported function name.

It would obviously be better if we just kept the exported function names
in one place, but I’m resigning myself to this hack for now so that I
can get things working without having to worry about how to generate
fresh identifiers for anonymous functions (to tie them into the
export AST node) and get those fresh identifiers into the identifier
context.

[0] https://webassembly.github.io/spec/core/text/modules.html#functions
[1] https://webassembly.github.io/spec/core/text/modules.html#exports