Initialise globals after module has been created
This is necessary because we need to call #evaluate_expression as part
of computing the initial value of each global, but we shouldn’t call
that without having a current module set. So first we create the module,
which itself doesn’t involve evaluating any expressions, and then
evaluate all of the global initialiser expressions.