Extract #function_at helper
We’ll need this soon to interpret the function index correctly. Because
function imports populate the functions index space of a module’s
initial identifier context [0], and “the indices of imports go before
the first index of any definition contained in the module itself” [1],
we need to check whether a function index is pointing at an imported
function versus a function defined in the module.

At the moment we don’t have any way to import functions so it’s always a
straightforward lookup in `current_module.functions`, but soon we’ll
introduce function imports and need a single place to make this decision
about where to look.

[0] https://webassembly.github.io/spec/core/text/modules.html#text-module
[1] https://webassembly.github.io/spec/core/syntax/modules.html#imports