Support imported functions in #function_at
At the moment I’ve decided to just return an Import instance instead of
a Function instance, which isn’t very good, but the only thing that
needs to deal with this is #invoke_function, so for now I can just make
sure it knows what to do with an Import. I’ll do that in the next
commit.

More generally we need a common representation of “invokable function”,
regardless of whether it’s defined in WebAssembly or in native code (as
per `spectest`). We don’t have this representation yet, but it could
probably just be a proc…? I’ll think about that later.