Support top-level imported `spectest` functions in #invoke_function
If the module has a top-level `(import "spectest" "print_32" (func
(param i32)))`, then we can “invoke” that by just popping its one
argument off the stack. In principle we could print that value out,
which is what `print_32` is supposed to do, but I don’t actually want to
see that in the test output, so we’ll do nothing instead.