Support inline imported `spectest` functions in #invoke_function
This is the same behaviour as in the previous commit, but for syntax
that looks like `(func (import "spectest" "print_i32") (param i32))`.
Ultimately we’ll need to unify these two approaches by having this
syntax generate an explicit import rather than simply storing that
information on the (otherwise unused) AST node for a function
definition, but that’s a future problem to solve once everything is
working.