Populate `module.functions` with both imported and defined functions during instantiation
This is a step in the direction of hiding the details of the AST from
the interpreter: the runtime representation of modules now contains a
list of function-like things (either Function or Import instances, the
latter of kind `:func`) in the correct order and with the correct
indexes. AST details are still leaking through but at least everything
has the right index now.