Explicitly skip export from binary module in float_literals.wast
This is the only time we want to (for now) ignore the failure to find an
exported function, because it’s only this one occurrence in a binary
module in float_literals.wast that causes a problem; elsewhere we’d like
to fail fast if we can’t find an export. Being stricter about exactly
which failure we’re prepared to ignore also makes the body of
the #find_function method simpler, to the point where we intend to
remove it entirely in the next commit.