Resolve top-level function imports during module instantiation
I haven’t thought about the consequences of this being any function
other than `print_i32`, because I think that’s the only function that’s
been imported in the tests so far. It seems reasonably likely that it
won’t work if one WebAssembly module imports a function from another,
because at the moment we have no notion of which module a function comes
from; its body will always execute in the current module, which won’t
have the other functions, globals, tables, memories etc that it’s
expecting to see.

But it works fine for `print_i32`, and that’s all I’m worried about at
the moment.