Resolve inline function imports during module instantiation
We’re still leaving non-imported functions alone because we don’t yet
have a nice runtime representation of functions — they continue to be
AST nodes for now. But when the function definition is actually an
import of `print_i32` from `spectest`, it’ll be represented by the
appropriate Proc instead.

None of this is particularly graceful, but it’s an improvement upon the
previous situation where function imports were being resolved at
invocation time.